Updated dimensionality reduction technique.
Validity matrix is randomized along the ground-string axis. This means the clusters in our data according to any patterns along this axis (the most important ones) will not be immediately visible to PCA. We will use t-sne to sort symbol planes so that they are well clustered by similarity.  Then we can use PCA on the organized data to better quantify boundaries of and importance of categories. (This process can be repeated many times with different initializations and perplexities to find the best PCs possible.)